requirements/base.txt
channels==2.1.1
channels-redis==2.1.1
Django==2.0.4
django-environ==0.4.4
django-model-utils==3.1.2
djangorestframework==3.8.2
psycopg2==2.7.4